+++ modules/tinymce/src/core/main/ts/util/Quirks.ts
@@ -35,7 +35,7 @@ const Quirks = (editor: Editor): Quirks => {
   const browser = Env.browser;
   const isGecko = browser.isFirefox();
   const isIE = browser.isIE() || browser.isEdge();
-  const isWebKit = browser.isChrome() || browser.isSafari();
+  const isWebKit = browser.isChromium() || browser.isSafari();
   const isiOS = Env.deviceType.isiPhone() || Env.deviceType.isiPad();
   const isMac = Env.os.isMacOS() || Env.os.isiOS();
   const mceInternalUrlPrefix = 'data:text/mce-internal,';