+++ modules/tinymce/src/plugins/charmap/test/ts/browser/CharmapSearchTest.ts
@@ -14,7 +14,7 @@ describe('browser.tinymce.plugins.charmap.SearchTest', () => {
   before(function () {
     // TODO: TINY-6905: Test is broken on Chromium Edge 86, so we need to investigate
     const platform = PlatformDetection.detect();
-    if (platform.browser.isChrome() && platform.os.isWindows()) {
+    if (platform.browser.isChromium() && platform.os.isWindows()) {
       this.skip();
     }
   });